refactor: 인증 페이지를 (auth) Route Group으로 이동 및 URL 구조 개선
