+++ src/app/(auth)/reset-password/step2/page.jsx
@@ -1,11 +1,11 @@
 'use client';
 import React, { useState, useEffect } from 'react';
 import { useRouter } from 'next/navigation';
-import { isValidPassword } from '../../../constants/regex';
-import axiosInstance from '../../../libs/api/instance';
-import Button from '../../../components/common/Button';
-import PrivacyPolicyFooter from '../../../components/common/PrivacyPolicyFooter';
-import FooterNav from '../../../components/common/FooterNav';
+import { isValidPassword } from '@constants/regex';
+import axiosInstance from '@lib/api/instance';
+import Button from '@components/common/Button';
+import PrivacyPolicyFooter from '@components/common/PrivacyPolicyFooter';
+import FooterNav from '@components/common/FooterNav';
 
 function BottomSafeSpacer({ height = 64 }) {
   return (
