+++ src/app/(auth)/sign-up/step1/customizedStepper.jsx
