+++ src/app/(auth)/sign-up/step1/page.jsx
@@ -1,11 +1,14 @@
 'use client';
 import React, { useEffect, useMemo, useRef, useState } from 'react';
+import { useRouter } from 'next/navigation';
 import Link from 'next/link';
-import { strictEmailRegex } from '../../../constants/regex';
-import axiosInstance from '../../../libs/api/instance';
-import Button from '../../../components/common/Button';
-import PrivacyPolicyFooter from '../../../components/common/PrivacyPolicyFooter';
-import FooterNav from '../../../components/common/FooterNav';
+import { strictEmailRegex } from '@constants/regex';
+import useSignupStore from '@stores/useSignupStore';
+import axiosInstance from '@lib/api/instance';
+import Button from '@components/common/Button';
+import PrivacyPolicyFooter from '@components/common/PrivacyPolicyFooter';
+import FooterNav from '@components/common/FooterNav';
+import CustomizedStepper from './customizedStepper';
 
 function BottomSafeSpacer({ height = 64 }) {
   return (
@@ -16,13 +19,13 @@ function BottomSafeSpacer({ height = 64 }) {
   );
 }
 
-export default function ResetPassWord1() {
+export default function SignUpStep1() {
+  const router = useRouter();
   const [email, setEmail] = useState('');
   const [number, setNumber] = useState('');
-
-  const [isCodeVerified, setIsCodeVerified] = useState(false);
   const [emailError, setEmailError] = useState('');
   const [numberError, setNumberError] = useState('');
+  const [isCodeVerified, setIsCodeVerified] = useState(false);
   const [codeVerificationMessage, setCodeVerificationMessage] = useState('');
 
   const [isSending, setIsSending] = useState(false);
@@ -31,6 +34,10 @@ export default function ResetPassWord1() {
   const [secondsLeft, setSecondsLeft] = useState(0);
   const [codeSent, setCodeSent] = useState(false);
   const timerRef = useRef(null);
+
+  const { setSignupField } = useSignupStore();
+
+  const MANUAL_URL = '/email-verification-manual';
 
   const commonCodeButtonClass =
     'inline-flex items-center justify-center w-[100px] h-10 ' +
@@ -95,7 +102,7 @@ export default function ResetPassWord1() {
     }
   };
 
-  const handleCodeInput = (value) => {
+  const handleNumberChange = (value) => {
     const v = value.replace(/\s/g, '');
     setNumber(v);
     setIsCodeVerified(false);
@@ -132,7 +139,6 @@ export default function ResetPassWord1() {
         email,
         code: number,
       });
-
       const ok =
         (res?.data?.verified ?? res?.data?.success) === true ||
         res?.status === 200;
@@ -156,15 +162,17 @@ export default function ResetPassWord1() {
   };
 
   const handleNext = () => {
-    sessionStorage.setItem('resetEmail', email);
+    setSignupField('email', email);
+    router.push('/login/sign-up-step2');
   };
+
+  const canVerify = codeSent && secondsLeft > 0 && /^[0-9]{6}$/.test(number);
 
   return (
     <div className="min-h-screen bg-gray-50 flex flex-col">
-      {/* 스피너 오버레이 */}
       {isSending && (
         <div
-          className="fixed inset-0 bg-black bg-opacity-50 z-50 flex items-center justify-center"
+          className="fixed inset-0 z-[9999] bg-black/40 backdrop-blur-sm md:backdrop-blur flex items-center justify-center"
           role="status"
           aria-live="polite"
         >
@@ -172,7 +180,7 @@ export default function ResetPassWord1() {
             <div
               className="w-8 h-8 border-4 border-[#788cff] border-t-transparent rounded-full animate-spin"
               aria-label="로딩 중"
-            ></div>
+            />
             <p className="text-sm text-gray-600 text-center">
               메일을 전송 중이예요. 시스템 환경에 따라 딜레이가 발생할 수
               있어요.
@@ -181,10 +189,14 @@ export default function ResetPassWord1() {
         </div>
       )}
 
-      <main className="flex-1 px-6 py-8">
+      <div className="flex-1 px-6 py-8">
         <div className="text-center space-y-3 mb-8">
-          <h1 className="text-2xl font-bold text-[#37352f]">비밀번호 재설정</h1>
-          <p className="text-[#73726e] text-sm">등록한 이메일로 찾기</p>
+          <h1 className="text-2xl font-bold text-[#37352f]">회원가입</h1>
+          <p className="text-[#73726e] text-sm">학교 이메일 인증하기</p>
+        </div>
+
+        <div className="mb-8">
+          <CustomizedStepper />
         </div>
 
         <div className="max-w-md mx-auto w-full space-y-6">
@@ -232,13 +244,13 @@ export default function ResetPassWord1() {
             <label className="block text-sm font-medium text-[#37352f]">
               인증번호
             </label>
-            <div className="flex items-center gap-2">
+            <div className="flex items-center gap-2 flex-nowrap">
               <div className="relative flex-1 min-w-0">
                 <StyledNumberInput
                   type="text"
                   id="number"
                   value={number}
-                  onChange={(e) => handleCodeInput(e.target.value)}
+                  onChange={(e) => handleNumberChange(e.target.value)}
                   placeholder="인증번호를 입력해주세요."
                   inputMode="numeric"
                   maxLength={6}
@@ -254,13 +266,7 @@ export default function ResetPassWord1() {
               <button
                 className={commonCodeButtonClass}
                 onClick={handleVerifyCode}
-                disabled={
-                  isVerifying ||
-                  !codeSent ||
-                  secondsLeft <= 0 ||
-                  !/^[0-9]{6}$/.test(number) ||
-                  isSending
-                }
+                disabled={!canVerify || isVerifying || isSending}
               >
                 {isVerifying ? '확인중' : '인증번호 확인'}
               </button>
@@ -270,29 +276,66 @@ export default function ResetPassWord1() {
             )}
             {codeVerificationMessage && (
               <p
-                className={`text-xs mt-1.5 ${
-                  isCodeVerified ? 'text-green-600' : 'text-red-500'
-                }`}
+                className={`text-xs mt-1.5 ${isCodeVerified ? 'text-green-600' : 'text-red-500'}`}
               >
                 {codeVerificationMessage}
               </p>
             )}
           </div>
 
-          <div className="mt-8">
-            <Link
-              href={isCodeVerified ? '/login/reset-password-step2' : '#'}
-              onClick={(e) => !isCodeVerified && e.preventDefault()}
-            >
-              <Button
-                onClick={handleNext}
-                disabled={!isCodeVerified}
-                text="다음으로"
-              />
-            </Link>
+          <div
+            className="rounded-lg border border-[#e9e9e7] bg-white p-4 sm:p-5 space-y-3"
+            aria-live="polite"
+          >
+            <div className="flex items-start gap-3">
+              <span
+                className="mt-0.5 inline-flex h-5 w-5 items-center justify-center rounded-full border border-[#d9d9d6] text-xs text-[#73726e]"
+                aria-hidden="true"
+              >
+                i
+              </span>
+              <div className="flex-1 space-y-2">
+                <h3 className="text-sm font-semibold text-[#37352f]">
+                  혹시 메일이 도착하지 않는다면?
+                </h3>
+                <p className="text-xs text-[#73726e] leading-relaxed">
+                  학교 계정으로 이메일이 발송되지 않는다면, 해당 계정이{' '}
+                  <b>2023년 11월 이전 발급된 계정</b>으로{' '}
+                  <b>명지대학교 측에서 메일을 이관하여 현재 사용정지된 상태</b>
+                  일 수 있습니다. 아래 <b>해결방법</b> 버튼을 클릭하여{' '}
+                  <b>메일 이관</b> 후 회원가입하시길 바랍니다.
+                </p>
+
+                <div className="pt-1">
+                  <Link
+                    href={MANUAL_URL}
+                    className="inline-flex items-center justify-center h-10 px-3 rounded-lg border border-[#788cff] text-[#788cff] hover:bg-[#788cff] hover:text-white text-sm font-medium transition-all duration-200"
+                    aria-label="이메일 인증 문제 해결 매뉴얼 새 창으로 열기"
+                  >
+                    해결방법
+                    <svg
+                      xmlns="http://www.w3.org/2000/svg"
+                      viewBox="0 0 20 20"
+                      fill="currentColor"
+                      className="ml-1 h-4 w-4"
+                      aria-hidden="true"
+                    >
+                      <path d="M12.293 2.293a1 1 0 011.414 0l4 4A1 1 0 0117 8h-3a1 1 0 110-2h.586L12 3.414V4a1 1 0 11-2 0V3a1 1 0 011-1h1.293z" />
+                      <path d="M3 5a2 2 0 012-2h4a1 1 0 110 2H5v10h10v-4a1 1 0 112 0v4a2 2 0 01-2 2H5a2 2 0 01-2-2V5z" />
+                    </svg>
+                  </Link>
+                </div>
+              </div>
+            </div>
           </div>
+
+          <Button
+            onClick={handleNext}
+            disabled={!isCodeVerified}
+            text="다음으로"
+          />
         </div>
-      </main>
+      </div>
 
       <PrivacyPolicyFooter />
       <BottomSafeSpacer height={64} />
@@ -302,28 +345,35 @@ export default function ResetPassWord1() {
 }
 
 const StyledInput = ({ value, className = '', ...props }) => {
-  return (
-    <input
-      className={`w-full px-4 py-3 bg-white rounded-lg border border-[#e9e9e7] text-sm placeholder:text-[#9b9998] focus:outline-none focus:border-[#788cff] focus:ring-2 focus:ring-[#788cff]/10 transition-all duration-200 ${className}`}
-      value={value}
-      {...props}
-    />
-  );
+  const base =
+    'w-full px-4 py-3 bg-white rounded-lg border border-[#e9e9e7] text-sm ' +
+    'placeholder:text-[#9b9998] focus:outline-none focus:border-[#788cff] ' +
+    'focus:ring-2 focus:ring-[#788cff]/10 transition-all duration-200';
+  return <input className={`${base} ${className}`} value={value} {...props} />;
 };
 
-const StyledEmailInput = ({ value, setEmail, disabled, ...props }) => {
-  const handleRemoveEmailValue = () => {
-    setEmail('');
-  };
-
+const StyledEmailInput = ({
+  value,
+  setEmail,
+  className = '',
+  disabled,
+  ...props
+}) => {
+  const handleRemoveEmailValue = () => setEmail('');
   return (
     <div className="relative">
-      <StyledInput {...props} value={value} disabled={disabled} />
+      <StyledInput
+        {...props}
+        value={value}
+        className={className}
+        disabled={disabled}
+      />
       {value && !disabled && (
         <button
           type="button"
           onClick={handleRemoveEmailValue}
           className="absolute right-3 top-1/2 -translate-y-1/2 p-1 hover:bg-gray-100 rounded-md transition-colors"
+          aria-label="이메일 입력 내용 지우기"
         >
           <img
             src="/static/icons/x_icon.svg"
