+++ src/app/(auth)/sign-up/step2/customizedStepper.jsx
