+++ src/app/(auth)/sign-up/step3/customizedStepper.jsx
