+++ src/app/(auth)/sign-up/step3/page.jsx
@@ -2,10 +2,10 @@
 import React, { useState, useEffect } from 'react';
 import { useRouter } from 'next/navigation';
 import axios from 'axios';
-import useSignupStore from '../../../stores/useSignupStore';
-import Button from '../../../components/common/Button';
-import PrivacyPolicyFooter from '../../../components/common/PrivacyPolicyFooter';
-import FooterNav from '../../../components/common/FooterNav';
+import useSignupStore from '@stores/useSignupStore';
+import Button from '@components/common/Button';
+import PrivacyPolicyFooter from '@components/common/PrivacyPolicyFooter';
+import FooterNav from '@components/common/FooterNav';
 import CustomizedStepper from './customizedStepper';
 
 function BottomSafeSpacer({ height = 64 }) {
