+++ src/app/(auth)/sign-up/step4/customizedStepper.jsx
